% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_simu.R
\name{all_metrics}
\alias{all_metrics}
\title{Compute all 5 metrics for a given FDR threshold}
\usage{
all_metrics(df, cutoff)
}
\arguments{
\item{df}{a data.frame, the output of the simulations}

\item{cutoff}{The FDR nominal level to control}
}
\description{
Compute all 5 metrics for a given FDR threshold
}
