% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_simu.R
\name{anayze_all}
\alias{anayze_all}
\title{Analyse dataset with two conditions}
\usage{
anayze_all(sce, shape = 2)
}
\arguments{
\item{sce}{The dataset, a SingleCellExperiment object}

\item{shape}{Number of lineages (2, 3 or 5)}
}
\description{
Analyse dataset with two conditions
}
