% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_simu.R
\name{anayze_all_per_cond}
\alias{anayze_all_per_cond}
\title{Analyse dataset with common and separate trajectories per condition}
\usage{
anayze_all_per_cond(sce, shape = 2)
}
\arguments{
\item{sce}{The dataset, a SingleCellExperiment object}

\item{shape}{Number of lineages (2, 3 or 5.)}
}
\description{
Analyse dataset with common and separate trajectories per condition
}
