% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_simu.R
\name{anayze_multiple_conditions}
\alias{anayze_multiple_conditions}
\title{Analyse dataset with three conditions}
\usage{
anayze_multiple_conditions(sce)
}
\arguments{
\item{sce}{The dataset, a SingleCellExperiment object}

\item{shape}{Number of lineages (2 or 3)}
}
\description{
Analyse dataset with three conditions
}
