% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_simu.R
\name{clean_results}
\alias{clean_results}
\title{Clean the output from the various simulation}
\usage{
clean_results(df)
}
\arguments{
\item{df}{a data.frame, the output of the simulations}

\item{nSim}{number of simulations for each condition}
}
\description{
Clean the output from the various simulation
}
