% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{complex}
\alias{complex}
\title{Results on the one bifurcation and three conditions datasets}
\format{
Results on 900 datasets
}
\source{
Output of the simulations
}
\usage{
data(complex)
}
\description{
This dataset contains all the results from the simulation on the datasets
with one bifurcation and three conditions
}
\keyword{datasets}
