% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{complex_sce}
\alias{complex_sce}
\title{A \link{\code{SingleCellExperiment}} example of a one bifurcation and three conditions dataset}
\format{
A \link{\code{SingleCellExperiment}} object
}
\source{
Created by running "fork_sce <- create_bifurcating_three_conditions_simu(multiplier = .5)"
}
\usage{
data(complex_sce)
}
\description{
Simulated dataset with one bifurcation and three conditions
}
\keyword{datasets}
