% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{condRes_monocle}
\alias{condRes_monocle}
\title{A \link[data.frame]{data.frame} object}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 10549 rows and 5 columns.
}
\source{
See the TGFB vignette to generate this data
}
\usage{
data(condRes_monocle)
}
\description{
This contains the output of running the \link[tradeSeq]{conditionTest} on
the \link{tgfb} dataset after running \link{monocle3}.
}
\details{
#' @references
Jos\'{e}  L.  McFaline-Figueroa,  Andrew  J.  Hill,  Xiaojie  Qiu,
Dana  Jackson,  Jay  Shendure,  and  Cole Trapnell.
*A pooled single-cell genetic screen identifies regulatory checkpoints in *
\emph{the continuum of the epithelial-to-mesenchymal transition.}
Nature Genetics, 51(9):1389–1398, sep 2019.  ISSN 15461718.
doi:  10.1038/s41588-019-0489-5.
}
\keyword{datasets}
