% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_simu.R
\name{create_5_lineages_simu}
\alias{create_5_lineages_simu}
\title{Create a bifurcating trajectory with five lineages and three conditions}
\usage{
create_5_lineages_simu(multiplier = 1, nSim = 200)
}
\arguments{
\item{multiplier}{Effect on the knockout}

\item{nSim}{number of simulations for each condition}
}
\description{
Create a bifurcating trajectory with five lineages and three conditions
}
