% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_simu.R
\name{create_bifurcating_three_conditions_simu}
\alias{create_bifurcating_three_conditions_simu}
\title{Create a bifurcating trajectory with three conditions}
\usage{
create_bifurcating_three_conditions_simu(multiplier = 1, nSim = 200)
}
\arguments{
\item{multiplier}{Effect on the knockout}

\item{nSim}{number of simulations for each condition}
}
\description{
Create a bifurcating trajectory with three conditions
}
