% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_simu.R
\name{create_consecutive_bifurcating_simu}
\alias{create_consecutive_bifurcating_simu}
\title{Create a trajectory with two bifurcations and two conditions}
\usage{
create_consecutive_bifurcating_simu(multiplier = 1, nSim = 200)
}
\arguments{
\item{multiplier}{Effect on the knockout}

\item{nSim}{number of simulations for each condition}
}
\description{
Create a trajectory with two bifurcations and two conditions
}
