% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{five_lins}
\alias{five_lins}
\title{Results on the five lineages datasets}
\format{
Results on 900 datasets
}
\source{
Output of the simulations
}
\usage{
data(five_lins)
}
\description{
This dataset contains all the results from the simulation on the datasets
with five lineages
}
\keyword{datasets}
