% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fivelin_sce}
\alias{fivelin_sce}
\title{A \link[SingleCellExperiment]{SingleCellExperiment} example of 5 lineages and
two conditions dataset}
\format{
A \link[SingleCellExperiment]{SingleCellExperiment} object
}
\source{
Created by running "fivelin_sce <- create_5_lineages_simu(multiplier = .5)"
}
\usage{
data(fivelin_sce)
}
\description{
Simulated dataset with 5 lineages and two conditions
}
\keyword{datasets}
