% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fork}
\alias{fork}
\title{A \link{\code{SingleCellExperiment}} object}
\format{
Results on 2000 datasets
}
\source{
Output of the simulations
}
\usage{
data(fork)
}
\description{
This dataset contains all the results from the simulation on the datasets
with one bifurcation and two conditions
}
\keyword{datasets}
