% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fork_sce}
\alias{fork_sce}
\title{A \link{\code{SingleCellExperiment}} example of a one bifurcation and two conditions dataset}
\format{
A \link{\code{SingleCellExperiment}} object
}
\source{
Created by running "fork_sce <- create_bifurcating_simu(multiplier = .5)"
}
\usage{
data(fork_sce)
}
\description{
Simulated dataset with one bifurcation and two conditions
}
\keyword{datasets}
