% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_fibrosis.R
\name{import_fibrosis}
\alias{import_fibrosis}
\title{Download and preprocess the raw dataset from .}
\usage{
import_fibrosis()
}
\value{
A \code{SingleCellExperiment} object
}
\description{
Download the dataset from GEO, filter, and create a
\code{SingleCellExperiment} object
}
\references{
Arun C. Habermann et al.
\emph{Single-cell RNA sequencing reveals profibrotic roles of distinct epithelial and mesenchymal lineages in pulmonary fibrosis}
Nature, 577(7790):421–425,jan  2020. ISSN  14764687. doi: 10.1038/s41586-019-1884-x
}
