% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_TCDD.R
\name{import_TCDD}
\alias{import_TCDD}
\title{Download and import the raw dataset from Nault et al.}
\usage{
import_TCDD()
}
\description{
Download the dataset from GEO, filter, and create a
\code{Seurat object}
}
\references{
Rance Nault,  Kelly A. Fader,  Sudin Bhattacharya,  and Tim R. Zacharewski.
\emph{Single-Nuclei RNA Sequencing Assessment of the Hepatic Effects of}
\itemize{
\item 2,3,7,8-Tetrachlorodibenzo-p-dioxin.CMGH*,
11(1):147–159, jan 2021.  ISSN 2352345X.
doi:  10.1016/j.jcmgh.2020.07.012.
}
}
