% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_backbone_modulenet_simplify}
\alias{plot_backbone_modulenet_simplify}
\title{Plot simulation dataset by condition and cluster}
\usage{
plot_backbone_modulenet_simplify(model)
}
\arguments{
\item{model}{Output of \code{\link{initialise_model}}}
}
\description{
Takes as input the low dimension representation and the
skeleton structure of WT and KO and plot them, colored by cluster
}
\details{
Code mostly copied from https://github.com/dynverse/dyngen/blob/master/R/plotting.R
}
