% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_example}
\alias{plot_example}
\title{Plot examples for figure 1}
\usage{
plot_example(sd_ko, mst_ko, sd_wt, mst_wt, title, colors)
}
\arguments{
\item{sd_ko}{Low dim rep for KO}

\item{mst_ko}{network structure for KO}

\item{sd_wt}{Low dim rep for WT}

\item{mst_wt}{network structure for WT}

\item{title}{Plot title}

\item{colors}{color scheme}
}
\description{
Takes as input the low dimension representation and the
skeleton structure of WT and KO and plot them.
}
