% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_reduced_dim}
\alias{plot_reduced_dim}
\title{Plot simulation dataset by condition and cluster}
\usage{
plot_reduced_dim(sce)
}
\arguments{
\item{sce}{A single cell experiment object, result of the simulation}
}
\description{
Takes as input the low dimension representation and the
skeleton structure of WT and KO and plot them, colored by cluster
}
