% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{step1_fail}
\alias{step1_fail}
\title{Results of the unstability test}
\format{
Results in data.frame format
}
\source{
Output of the simulations
}
\usage{
data(step1_fail)
}
\description{
This dataset contains all the results from the step1_fail test
}
\keyword{datasets}
