% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tcdd}
\alias{tcdd}
\title{A \link[SingleCellExperiment]{SingleCellExperiment} object}
\format{
An object of class \code{SingleCellExperiment} with 8027 rows and 9951 columns.
}
\source{
See the TCDD vignette to generate this data
}
\usage{
data(tcdd)
}
\description{
This dataset contains all the information used for the TCDD case studies
}
\references{
Rance Nault,  Kelly A. Fader,  Sudin Bhattacharya,  and Tim R. Zacharewski.
\emph{Single-Nuclei RNA Sequencing Assessment of the Hepatic Effects of}
\itemize{
\item 2,3,7,8-Tetrachlorodibenzo-p-dioxin.CMGH*,
11(1):147–159, jan 2021.  ISSN 2352345X.
doi:  10.1016/j.jcmgh.2020.07.012.
}
}
\keyword{datasets}
