% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tree}
\alias{tree}
\title{Results on the two bifurcations and two conditions datasets}
\format{
Results on 2000 datasets
}
\source{
Output of the simulations
}
\usage{
data(tree)
}
\description{
This dataset contains all the results from the simulation on the datasets
with two bifurcations and two conditions
}
\keyword{datasets}
