% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{unstability}
\alias{unstability}
\title{Results of the unstability test}
\format{
Results in data.frame format
}
\source{
Output of the simulations
}
\usage{
data(unstability)
}
\description{
This dataset contains all the results from the unstability test
}
\keyword{datasets}
