% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fibrosis}
\alias{fibrosis}
\title{A \link[SingleCellExperiment]{SingleCellExperiment} object}
\format{
An object of class \code{SingleCellExperiment} with 10100 rows and 14462 columns.
}
\source{
See the fibrosis vignette to generate this data
}
\usage{
data(fibrosis)
}
\description{
This dataset contains all the information used for the fibrosis case study
}
\references{
Arun  C.  Habermann,  Austin  J.  Gutierrez,  Linh  T.  Bui,
Stephanie  L. Yahn,  Nichelle  I.  Winters, Carla L. Calvi, Lance Peter,
Mei I. Chung, Chase J. Taylor, Christopher Jetter, Latha Raju, Jamie Roberson,
Guixiao Ding, Lori Wood, Jennifer M.S. Sucre, Bradley W. Richmond, Ana P. Serezani,
Wyatt J. McDonnell,  Simon B. Mallal,  Matthew J. Bacchetta,  James E. Loyd,
Ciara M. Shaver,Lorraine B. Ware,  Ross Bremner,  Rajat Walia,
Timothy S. Blackwell,  Nicholas E. Banovich,  and Jonathan A. Kropski.
\emph{Single-cell rna sequencing reveals profibrotic roles of distinct epithelial and mesenchymal  lineages  in  pulmonary  fibrosis}
Science  Advances,  6,  2020.   ISSN  23752548.   doi:10.1126/sciadv.aba1972.
}
\keyword{datasets}
