% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_reduced_dim_together}
\alias{plot_reduced_dim_together}
\title{Plot simulation dataset by condition}
\usage{
plot_reduced_dim_together(sce, sample = 1)
}
\arguments{
\item{sce}{A single cell experiment object, result of the simulation}

\item{sample}{Fraction of the data to plot. Default to 1.}
}
\description{
Takes as input the low dimension representation and the
skeleton structure of WT and KO and plot them, colored by condition.
}
