% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tree_sce}
\alias{tree_sce}
\title{A \link{\code{SingleCellExperiment}} example of a two bifurcations and two conditions dataset}
\format{
A \link{\code{SingleCellExperiment}} object
}
\source{
Created by running "tree_sce <- create_consecutive_bifurcating_simu(multiplier = .5)"
}
\usage{
data(tree_sce)
}
\description{
Simulated dataset with two bifurcations and two conditions
}
\keyword{datasets}
